Move thai orchid & best tokenization to the tokenization specific directory
